Modularize vitest setup into separate files
Split vitestSetup.ts into a thin entry point that imports modular setup files.

New structure:
- setup/polyfills.ts: Blob, ResizeObserver polyfills
- setup/domStubs.ts: Radix UI pointer capture stubs
- setup/mocks.ts: useProxyLatency mock
- setup/msw.ts: MSW server lifecycle
