+++ site/test/setup/domStubs.ts
@@ -1,0 +1,6 @@
+// Pointer capture stubs required for Radix UI in JSDOM.
+globalThis.HTMLElement.prototype.hasPointerCapture = vi
+	.fn()
+	.mockReturnValue(false);
+globalThis.HTMLElement.prototype.setPointerCapture = vi.fn();
+globalThis.HTMLElement.prototype.releasePointerCapture = vi.fn();
