+++ site/test/setup/msw.ts
@@ -1,0 +1,11 @@
+import { server } from "testHelpers/server";
+import { cleanup } from "@testing-library/react";
+
+// MSW server lifecycle
+beforeAll(() => server.listen({ onUnhandledRequest: "warn" }));
+afterEach(() => {
+	cleanup();
+	server.resetHandlers();
+	vi.clearAllMocks();
+});
+afterAll(() => server.close());
